requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
sqlalchemy
alembic
aiosqlite
flask
gunicorn
flask-sqlalchemy
flask-cors